No need to have variable names in phpdoc comments
"@var <type> <name>" does not work for variables. There has to be
a "$" before the name for it to show up correctly in the docs with
phpdoc. I think this was copy-pasted from the Ruby client.

The variable name can also be omitted when there is a comment above
each individual variable, which it is in this case.

see http://www.phpdoc.org/docs/latest/references/phpdoc/tags/var.html