Update phpunit to support newer PHP versions
We started testing on newer PHP versions in ee6f58e, and 7.3 is the
lowest version we test on.

PHPUnit 9 is compatible with >= 7.3: https://phpunit.de/supported-versions.html

---

Got the following errors related to the new version of PHPUnit, so I
went ahead and fixed those:

    Uncaught Error: Class 'PHPUnit_Framework_TestCase' not found in
        twingly-search-api-php/tests/ClientTest.php:9
    Declaration of Twingly\Tests\ClientTest::setUp() must be compatible
        with PHPUnit\Framework\TestCase::setUp(): void in
        twingly-search-api-php/tests/ClientTest.php on line 17

---

Because some tests overloaded the constructor for the `TestCase` class, the
following cryptic error was shown in the output:

    array_merge(): Expected parameter 1 to be an array, null given

To get rid of the error I switched to using the `setUpBeforeClass` function
instead. However, in the `QueryTest` that wasn't possible, since we use `$this`,
so I just moved the setup into the `setUp` function, which gets run before
each test.

---

Switched from `assertContains` to `assertStringContainsString` because of
the following error:

    TypeError: Argument 2 passed to PHPUnit\Framework\Assert::assertContains()
               must be iterable, string given